feat: add duplicator for testing
We want to be able to ensure the new Tablespoon software is handling the same
requests that the existing TSP software is. To support that, we'll run
`duplicator`, which sends the HTTP requests to the existing server as well as
the new server, but ignores the responses from the new server. This will
allow us to ensure the new software is handling the same messages the current
server is.