feat: PMPPMultiplex for multiple packets over a single transport
Since all the BTD messages will go over a single TCP connection, we need to
wrap that single connection in a GenServer that they all can share. This
way, we keep our existing Transport infrastructure so the intersections
themselves don't need to care that the messages are going over a single connection.