cleanup(Intersection): use handle_continue instead of a message
handle_continue/2 was added as a callback in OTP 21. It allows use to enforce
that the connect call is made before any other messages are handled. The
immediate send of a message in init/2 kind of works for this (which is what
we were doing before), but handle_continue is a better implementation.