feat(PMPPMultiplex): use id_mfa to distinguish requests
The new BTD code should be supporting sending responses out of order,
so we need to be able to distinguish the responses. In order to avoid
fully parsing the data twice, we defer to a provided MFA to calculate
the request ID and use that to match the response to the request.