fix(NTCIP1211Extended): decode_id/1 better handles invalid packets
During testing, we sometimes received packets which were too short. Since the
built-in `:snmp_pdus` module doesn't have support for partial parsing of a
packet, we implement enough of the ASN.1 parsing logic to pull a request ID
out of a packet, even if it's otherwise invalid.

This allows PMPPMultiplex to route the invalid packet to correct
Communicator (hopefully) without needing to terminate the connection
entirely.