test: slight refactor in PMPPMultiplexTest
One of our tests was avoiding timeouts by sending a message with a
reference, and then waiting for that reference to come back. However, we
could still lose the race and get the reference message before the message
back from the client process.

Now, we loop waiting for a message, and if it's a message the transport
ignores we drop it on the floor and move along. There's still a 5s timeout,
but hopefully we never hit it.