feat(Modem): re-connect if there are stale queries
Sometimes, the modems do not send a reply in a timely fashion, or perhaps
drop a reply. Since we can't tell that a reply is for a particular request,
this means that the replies get more and more stale.

Instead, we check for outstanding queries that are more than 30s old: if so,
we mark all the queries as failed and re-connect.