refactor(Modem): `connected?` -> `connection_state`
In order to support new Picocom modems, we'll need to track more connection
states. We expand the state of the connection to be one of a few different
atoms: right now, `:not_connected`, `:awaiting_ok` and `:connected`.

This also uncovered an issue with our property test, where we weren't
actually running through the reconnection during the test.