feat(Modem): cancel requests in the background if we don't see one
With the Modem protocol, we can only have the modem in one of two states:
on and off. If we get two requests and a cancel, we don't want the cancel to
take effect, as there's still a vehicle waiting to pass through the
intersection.

However, if that second vehicle doesn't send us a cancel message, we can get
stuck in the on position. This change addresses that, by keeping a queue of
vehicles who have requested that the intersection stay on. If a timeout
expires and the vehicle has not sent a cancel message, we'll do it in the
background and log a warning.