feat: support loading `runtime.exs` from an envvar
We need to be able to configure the intersections, as well as other
configuration, outside of this repo. Previously, this was split between a
`prod.secret.exs` which was loaded at compile time, and `INTERSECTIONS_JSON`
loaded at runtime.

Now, we provide a `CONFIG_RUNTIME_EXS` environment variable, which is written
to `runtime.exs` before the application starts. This provides us one
file/secret to configure.