fix(PMPPMultiplex): don't crash when closing a closed transport
Once one Intersection closes the connection, the others also become closed as
they are notified of the closure. If the other Intersections are also
closing (say during shutdown), this results in duplicate calls to
`close/1`. Since the connection has already been closed, we don't need to do
anything and can return the transport as-is.